Refactor reusable client type modules
