+++ custom_components/dreame_lawn_mower/dreame_lawn_mower_client/const.py
@@ -1,5 +1,5 @@
 from typing import Final
-from .types import (
+from .device_types import (
     DreameMowerChargingStatus,
     DreameMowerTaskStatus,
     DreameMowerState,
