+++ custom_components/dreame_lawn_mower/dreame_lawn_mower_client/map.py
@@ -36,7 +36,7 @@ from threading import Timer
 from .resources import *
 from .protocol import DreameMowerProtocol
 from .exceptions import DeviceUpdateFailedException
-from .types import (
+from .device_types import (
     PIID,
     DIID,
     DreameMowerProperty,
@@ -46,42 +46,46 @@ from .types import (
     RobotType,
     CleansetType,
     ObstacleType,
-    Obstacle,
     FurnitureType,
-    Furniture,
     PathType,
-    Point,
-    Coordinate,
-    MapDataPartial,
-    MapData,
-    MapFrameType,
-    MapPixelType,
-    Path,
-    Area,
-    Wall,
-    Segment,
-    StartupMethod,
-    CleanupMethod,
-    TaskEndType,
-    RecoveryMapType,
     ObstacleIgnoreStatus,
-    MapImageDimensions,
-    MapRendererLayer,
-    MapRendererColorScheme,
-    MapRendererConfig,
-    MapRendererData,
-    MapRendererResources,
-    RecoveryMapInfo,
-    MAP_COLOR_SCHEME_LIST,
-    MAP_ICON_SET_LIST,
     SEGMENT_TYPE_CODE_TO_NAME,
     SEGMENT_TYPE_CODE_TO_HA_ICON,
     FURNITURE_TYPE_TO_DIMENSIONS,
     FURNITURE_V2_TYPE_TO_DIMENSIONS,
+)
+from .map_renderer_types import (
     ALine,
-    CLine,
-    Paths,
+    MAP_COLOR_SCHEME_LIST,
+    MAP_ICON_SET_LIST,
     Angle,
+    CLine,
+    MapRendererColorScheme,
+    MapRendererConfig,
+    MapRendererData,
+    MapRendererLayer,
+    MapRendererResources,
+    Paths,
+)
+from .map_types import (
+    Area,
+    CleanupMethod,
+    Coordinate,
+    Furniture,
+    MapData,
+    MapDataPartial,
+    MapFrameType,
+    MapImageDimensions,
+    MapPixelType,
+    Obstacle,
+    Path,
+    Point,
+    RecoveryMapInfo,
+    RecoveryMapType,
+    Segment,
+    StartupMethod,
+    TaskEndType,
+    Wall,
 )
 from .const import (
     MAP_PARAMETER_NAME,
